% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{mclm_question}
\alias{mclm_question}
\alias{mclm_question_num}
\alias{mclm_question_text}
\title{Learnr question wrappers with desired defaults}
\usage{
mclm_question(...)

mclm_question_num(q, a, min = 5000, step = 100, tolerance = 0.001)

mclm_question_text(...)
}
\arguments{
\item{...}{A question and a series of [learnr::answer()].}

\item{q}{A question.}

\item{a}{The correct answer.}
}
\value{
A [learnr::question()] with desired defaults.
}
\description{
These are wrappers for [learnr::question()] and related functions that
always allow retries, give answers in a random order, and use
[learnr::random_praise()] and encouragement for the reactions.
To be used in tutorials.
}
\keyword{internal}
